+++ docs/changelog.rst
@@ -93,6 +93,8 @@ Detailed list of changes
 
 - Fix background image flashing when closing a tab (:iss:`7999`)
 
+- When running a kitten that modifies the kitty config file if no config file exists create a commented out default config file and then modify it (:iss:`7991`)
+
 0.36.4 [2024-09-27]
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~